Use .aar instead of external module for LibsUtils
This way it is much easier to manage and use with multiple projects. Previously any major incompatible change to library code was braking projects it was used in. Now if you want to update the library, you have to explicitly replace the .aar file and do all updates in one go