Rename dagger modules to factories
This app uses modular system for activities, fragments and other classes and names of this modules are clashing with dagger modules, so one of them has to be renamed. I couldn't find a good way to rename the activity etc. modules, but dagger modules can easily be renamed to factories and still retain their meaning and purpose